sample_storage: split backend.py into a driver/ package, one class per file
backend.py held the driver and all three capability backends. Split it into a
driver/ subpackage:

- driver/protocol.py   - wire tokens + parse_kv (shared helpers)
- driver/automated_retrieval.py
- driver/temperature.py
- driver/humidity.py
- driver/driver.py      - HighResSampleStorageDriver (owns the three backends)

Capability backends reference the driver via a TYPE_CHECKING import to avoid a
runtime import cycle; the driver/ __init__ re-exports the four classes so the
public import paths are unchanged.
